adds castles
