+++ Application/src/visuals/Coordinates.java
@@ -1,0 +1,19 @@
+package visuals;
+
+class Coordinates {
+    private int x;
+    private int y;
+
+    Coordinates(int x, int y) {
+        this.x = x;
+        this.y = y;
+    }
+
+    public int getX() {
+        return x;
+    }
+
+    public int getY() {
+        return y;
+    }
+}
